requirements.txt
opencv-python==4.7.0.72
tkinter==0.1.0
pandas==1.5.3
numpy==1.23.5
matplotlib==3.6.2
pillow==9.4.0
ffmpeg-python==4.4.0